Feat:스크랩 상태 로직 변경
