+++ app/feature/place/components/PlaceDetail/PlaceDetailTopBar.tsx
@@ -1,119 +1,43 @@
 "use client";
 
-import ScrapLine from "@common/assets/icons/scrap/ScrapLine";
-import ScrapFill from "@common/assets/icons/scrap/ScrapFill";
 import LinkIcon from "@common/assets/icons/link/LinkIcon";
-import { TopBarProps } from "../../../../common/components/ui/topBar/type";
 import BasicTopBar from "../../../../common/components/ui/topBar/BasicTopBar/BasicTopBar";
 import { copyLink } from "@common/utils/text/copy-link";
-import { getSession } from "@common/utils/session/getSession";
-import DeleteSpaceScrap from "@feature/place/queries/deleteScrapSpace";
-import PostSpaceScrap from "@feature/place/queries/postSpaceScrap";
 import { usePathname } from "next/navigation";
 import MapIcon from "@common/assets/icons/map/map";
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import Map from "@feature/map/components/Map/Map";
-import SaveModal from "@feature/record/components/Modal/SaveModal";
-import Toast from "@common/components/ui/toast/Toast";
-import revalidateScrapSpace from "@feature/place/actions/revalidateScrapSpace";
-import revalidateHomeRecommend from "@feature/place/actions/revalidateHomeRecommend";
-import revalidatePlaceDetailById from "@feature/place/actions/revalidatePlaceDetailById";
+
+interface PlaceDetailTopBar {
+  type: string;
+  address: string;
+  name: string;
+  imgUrl: string;
+  purpose: string[];
+  className: string;
+}
 
 export default function PlaceDetailTopBar({
-  id,
-  isScraped,
   type,
   address,
   name,
   imgUrl,
   purpose,
   className,
-}: Partial<TopBarProps> & {
-  type: string;
-  address: string;
-  name: string;
-  isScraped: boolean;
-  id: number;
-  imgUrl: string;
-  purpose: string[];
-}) {
+}: PlaceDetailTopBar) {
   const pathname = usePathname();
   const [mapOpen, setMapOpen] = useState(false);
-  const [openSaveModal, setOpenSaveModal] = useState(false);
-  const [scrapState, setScrapState] = useState<boolean>(isScraped);
-  const [openScrapToast, setOpenScrapToast] = useState(false);
-  const [toastText, setToastText] = useState("");
   const handleMapOpen = (state: boolean) => {
     setMapOpen(state);
-  };
-  const handleScrapClick = async (
-    e: React.MouseEvent<SVGSVGElement, MouseEvent>
-  ) => {
-    e.preventDefault();
-    const auth_info = await getSession();
-    const token = auth_info?.data?.accessToken;
-    if (!token) {
-      location.replace("/login");
-    } else {
-      if (scrapState) {
-        const res = await DeleteSpaceScrap(id as number);
-        if (res.status === 200) {
-          setOpenScrapToast(true);
-          setToastText(`스크랩이 해제되었습니다`);
-          setScrapState && setScrapState(false);
-          revalidateScrapSpace();
-          revalidateHomeRecommend();
-          revalidatePlaceDetailById(id);
-        } else {
-          alert("에러가 발생했습니다!");
-          return;
-        }
-      } else {
-        const res = await PostSpaceScrap(id as number);
-        if (res.status === 200) {
-          setOpenScrapToast(true);
-          setToastText(`스크랩 되었습니다`);
-          setScrapState && setScrapState(true);
-          setOpenSaveModal && setOpenSaveModal(true);
-          revalidateScrapSpace();
-          revalidateHomeRecommend();
-          revalidatePlaceDetailById(id);
-        } else {
-          alert("에러가 발생했습니다!");
-          return;
-        }
-      }
-    }
   };
   const handleCopyLinkClick = () => {
     copyLink(pathname);
   };
-  useEffect(() => {
-    let timeoutId: NodeJS.Timeout;
-    if (openScrapToast) {
-      timeoutId = setTimeout(() => {
-        setOpenScrapToast(false);
-      }, 1000);
-    }
-    return () => {
-      clearTimeout(timeoutId);
-    };
-  }, [openScrapToast]);
   return (
     <>
       <BasicTopBar className={className}>
         <div className="w-full flex justify-end items-center relative">
-          <MapIcon
-            color="white"
-            className="mr-[1.2rem]"
-            onClick={() => handleMapOpen(true)}
-          />
-          {scrapState ? (
-            <ScrapFill color="white" onClick={handleScrapClick} />
-          ) : (
-            <ScrapLine color="white" onClick={handleScrapClick} />
-          )}
-
+          <MapIcon color="white" onClick={() => handleMapOpen(true)} />
           <LinkIcon className="ml-[0.8rem]" onClick={handleCopyLinkClick} />
         </div>
       </BasicTopBar>
@@ -124,10 +48,6 @@ export default function PlaceDetailTopBar({
           className="fixed top-[7rem] z-10"
         />
       )}
-      {openSaveModal && (
-        <SaveModal spaceId={id} handleModalFn={setOpenSaveModal} />
-      )}
-      <Toast open={openScrapToast} text={toastText} />
     </>
   );
 }
