+++ app/feature/place/components/PlaceScrapList/PlaceScrapList.tsx
@@ -5,7 +5,7 @@ const PlaceInfoCard = dynamic(() => import("../PlaceInfoCard/PlaceInfoCard"));
 export default async function PlaceScrapList() {
   const scrappedPlace = await GetPlaceScrapped();
   return (
-    <div className="px-[2rem] gap-[0.8rem]">
+    <div className="flex px-[2rem] gap-[0.8rem]">
       {scrappedPlace?.length === 0 && (
         <div className="flex items-center justify-center w-full h-[18rem]">
           <p className="body1-medium text-text-gray-8">
